moved app bar files to their own folder
