+++ src/components/App/Home.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Grid, CssBaseline, useMediaQuery } from '@material-ui/core';
 import Calendar from '../Calendar/ScheduleCalendar';
-import Bar from './CustomAppBar';
+import Bar from '../AppBar/CustomAppBar';
 import DesktopTabs from '../CoursePane/DesktopTabs';
-import NotificationSnackbar from './NotificationSnackbar';
+import NotificationSnackbar from '../AppBar/NotificationSnackbar';
 import { MuiPickersUtilsProvider } from '@material-ui/pickers';
 import MobileHome from './MobileHome';
 import DateFnsUtils from '@date-io/date-fns';
